feat(bienes): agregar historial de dependencias de bienes
- Migración para crear la tabla `historial_dependencias_bienes`
- Seeder con datos de prueba
- Modelo Eloquent `HistorialDependenciaBien` para gestionar el historial
